refactor(auth): rename identity-module to auth-module
