+++ backend/src/api/private/auth/ldap/ldap.controller.ts
@@ -13,12 +13,12 @@ import {
 } from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 
+import { IdentityService } from '../../../../auth/identity.service';
+import { LdapLoginDto } from '../../../../auth/ldap/ldap-login.dto';
+import { LdapService } from '../../../../auth/ldap/ldap.service';
+import { ProviderType } from '../../../../auth/provider-type.enum';
+import { RequestWithSession } from '../../../../auth/session.guard';
 import { NotInDBError } from '../../../../errors/errors';
-import { IdentityService } from '../../../../identity/identity.service';
-import { LdapLoginDto } from '../../../../identity/ldap/ldap-login.dto';
-import { LdapService } from '../../../../identity/ldap/ldap.service';
-import { ProviderType } from '../../../../identity/provider-type.enum';
-import { RequestWithSession } from '../../../../identity/session.guard';
 import { ConsoleLoggerService } from '../../../../logger/console-logger.service';
 import { UsersService } from '../../../../users/users.service';
 import { makeUsernameLowercase } from '../../../../utils/username';
