+++ backend/src/api/private/media/media.controller.ts
@@ -19,8 +19,8 @@ import { FileInterceptor } from '@nestjs/platform-express';
 import { ApiBody, ApiConsumes, ApiHeader, ApiTags } from '@nestjs/swagger';
 import { Response } from 'express';
 
+import { SessionGuard } from '../../../auth/session.guard';
 import { PermissionError } from '../../../errors/errors';
-import { SessionGuard } from '../../../identity/session.guard';
 import { ConsoleLoggerService } from '../../../logger/console-logger.service';
 import { MediaUploadDto } from '../../../media/media-upload.dto';
 import { MediaService } from '../../../media/media.service';
