+++ backend/src/auth/pending-user-confirmation.dto.ts
