+++ backend/src/frontend-config/frontend-config.dto.ts
@@ -15,8 +15,8 @@ import {
 } from 'class-validator';
 import { URL } from 'url';
 
+import { ProviderType } from '../auth/provider-type.enum';
 import { GuestAccess } from '../config/guest_access.enum';
-import { ProviderType } from '../identity/provider-type.enum';
 import { ServerVersion } from '../monitoring/server-status.dto';
 import { BaseDto } from '../utils/base.dto.';
