+++ backend/src/seed.ts
@@ -6,11 +6,11 @@
 import { DataSource } from 'typeorm';
 
 import { ApiToken } from './api-token/api-token.entity';
+import { Identity } from './auth/identity.entity';
+import { ProviderType } from './auth/provider-type.enum';
 import { Author } from './authors/author.entity';
 import { Group } from './groups/group.entity';
 import { HistoryEntry } from './history/history-entry.entity';
-import { Identity } from './identity/identity.entity';
-import { ProviderType } from './identity/provider-type.enum';
 import { MediaUpload } from './media/media-upload.entity';
 import { Alias } from './notes/alias.entity';
 import { Note } from './notes/note.entity';
