+++ backend/src/users/users.module.ts
@@ -1,12 +1,12 @@
 /*
- * SPDX-FileCopyrightText: 2021 The HedgeDoc developers (see AUTHORS file)
+ * SPDX-FileCopyrightText: 2024 The HedgeDoc developers (see AUTHORS file)
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 
-import { Identity } from '../identity/identity.entity';
+import { Identity } from '../auth/identity.entity';
 import { LoggerModule } from '../logger/logger.module';
 import { Session } from '../sessions/session.entity';
 import { User } from './user.entity';
