+++ backend/test/private-api/auth.e2e-spec.ts
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2023 The HedgeDoc developers (see AUTHORS file)
+ * SPDX-FileCopyrightText: 2024 The HedgeDoc developers (see AUTHORS file)
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
@@ -10,10 +10,10 @@
 */
 import request from 'supertest';
 
+import { LoginDto } from '../../src/auth/local/login.dto';
+import { RegisterDto } from '../../src/auth/local/register.dto';
+import { UpdatePasswordDto } from '../../src/auth/local/update-password.dto';
 import { NotInDBError } from '../../src/errors/errors';
-import { LoginDto } from '../../src/identity/local/login.dto';
-import { RegisterDto } from '../../src/identity/local/register.dto';
-import { UpdatePasswordDto } from '../../src/identity/local/update-password.dto';
 import { UserRelationEnum } from '../../src/users/user-relation.enum';
 import { checkPassword } from '../../src/utils/password';
 import { Username } from '../../src/utils/username';
