+++ backend/test/private-api/groups.e2e-spec.ts
@@ -1,14 +1,14 @@
 /*
- * SPDX-FileCopyrightText: 2022 The HedgeDoc developers (see AUTHORS file)
+ * SPDX-FileCopyrightText: 2024 The HedgeDoc developers (see AUTHORS file)
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 import request from 'supertest';
 
+import { LoginDto } from '../../src/auth/local/login.dto';
 import { GuestAccess } from '../../src/config/guest_access.enum';
 import { createDefaultMockNoteConfig } from '../../src/config/mock/note.config.mock';
 import { NoteConfig } from '../../src/config/note.config';
-import { LoginDto } from '../../src/identity/local/login.dto';
 import {
   password1,
   TestSetup,
